+++ packages/react-utils/jest.config.ts
@@ -1,0 +1,6 @@
+/* eslint-disable */
+export default {
+  displayName: '@myraui-org/react-utils',
+  preset: '../../jest.preset.js',
+  coverageDirectory: '../../coverage/packages/react-utils',
+};