+++ packages/storybook/.storybook/theme.ts
@@ -3,5 +3,5 @@ import { create } from '@storybook/theming/create';
 export default create({
   base: 'dark',
   brandTitle: 'Myra UI',
-  brandUrl: 'https://github.com/myraui/myraui',
+  brandUrl: 'https://github.com/myraui-org/myraui',
 });