+++ packages/utilities/shared-utils/jest.config.ts
@@ -1,6 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: '@myraui/shared-utils',
-  preset: '../../../jest.preset.js',
-  coverageDirectory: '../../../coverage/packages/utilities/shared-utils',
-};